website_streamlit/requirements.txt
streamlit>=1.28.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
plotly>=5.15.0
google-auth>=2.17.0